General Improvement + YAMCAN Bringup Yamcan - Remove some unused nodes (eng) - Build out generation of MessagePack .c and .h - Update SConscript to trigger generation - Update CANIO-tx.c to support generated files - NetworkGen tool split between builder and node generator. The builder parses and caches the network while the generator retreives the cached files and then generates the code for a given node. Build system modified to build multiple targets, or a given platform at once. - Duplicate nodes support in yamcan - Added message RX functionality, addes messages definition to YAMCAN. Started integration into VC's - Timeout and counter message validation - Added transmit conditions and transmit condition stubbing - Implemented signedness into YAMCAN, fixed tx nd rx bit stuffing critical boundaries - Endiannes pack and unpack support - CANRX HW changes for ID list - Adding DiscreteValues to CAN file generation
General
- BMSW in yamcan
- BMSB in yamcan
- BMSB current sense and transmission to yamcan
- Added BMSB Contactor state to yamcan
- BMSW -Wconversion and code cleanup
- Added helper function for BMSW reading to BMSB. To be improved
- Added platform support. Added platform and generic build directory. Update all component scripts. Cache data at the network definition

+++ components/bms_worker/include/Module.h
@@ -11,6 +11,7 @@
 
 // System Includes
 #include "stdint.h"
+#include "FloatTypes.h"
 
 // Other Includes
 #include "ModuleDesc.h"
@@ -55,10 +56,10 @@ typedef enum
 
 typedef struct
 {
-    uint64_t total_runtime;
-    uint32_t timeslice_runtime;
-    uint8_t  total_percentage;
-    uint8_t  timeslice_percentage;
+    uint64_t  total_runtime;
+    uint64_t  timeslice_runtime;
+    float32_t  total_percentage;
+    float32_t  timeslice_percentage;
 } Module_taskStats_S;
 
 /******************************************************************************
